+++ Package.swift
@@ -4,7 +4,7 @@ import PackageDescription
 let package = Package(
     name: "FuturaeKit",
     platforms: [
-        .iOS(.v10)
+        .iOS(.v11)
     ],
     products: [
         .library(name: "FuturaeKit", targets: ["FuturaeKit"])
@@ -12,8 +12,8 @@ let package = Package(
     targets: [
         .binaryTarget(
           name: "FuturaeKit",
-          url: "https://artifactory.futurae.com/artifactory/futurae-ios/FuturaeKit-v2.3.1.xcframework.zip",
-          checksum: "35ab1511a793a6444674e5307958337ccdf961f665db8efb21701eea3894cb1a"
+          url: "https://github.com/Futurae-Technologies/ios-sdk/releases/download/v2.3.2/FuturaeKit-v2.3.2.xcframework.zip",
+          checksum: "61cd0d2f2fdab8c6da3d831f999154f97c77f1a891681061ba1ec1be5c7f9048"
         )
     ]
 )